Moved event source to home page only
